refactor(PapoleIcons): reorganize icon exports and remove unused SVG files
