+++ src/components/PapoleIcons/MicIconPapole.tsx
@@ -1,0 +1,27 @@
+import type { FC } from 'react';
+import type { PapoleIconProps } from './types';
+
+/**
+ * Microfone (ligado) — tema Papolê. Cor marrom `#604903` (fixa da arte).
+ * `aria-hidden` por ser decorativo; quem usa em botão deve prover o rótulo.
+ */
+export const MicIconPapole: FC<PapoleIconProps> = ({
+  size = 24,
+  className,
+}) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    className={className}
+    aria-hidden="true"
+  >
+    <path
+      d="M19.0899 9.64246C19.2722 9.12736 19.8996 8.87431 20.3966 9.09265C21.0667 9.38732 21.0012 10.0389 20.9688 10.6317C20.902 11.8547 20.6668 12.9383 20.1905 14.0663C19.8435 14.8412 19.4182 15.5484 18.8468 16.1815C17.3308 17.8298 15.2311 18.8217 12.9952 18.9471L13.0011 20.2665C13.0021 20.5743 13.0211 20.8906 12.9669 21.1962C12.9158 21.6151 12.4185 22.0384 11.9796 21.996C10.9121 21.8926 11.0195 20.9985 11.0206 20.3046L11.0235 18.955C6.82821 18.5619 3.81573 15.9926 3.14951 11.745C3.07466 11.2676 2.95457 10.2389 3.04502 9.78601C3.07111 9.6531 3.12649 9.52768 3.20713 9.41883C3.38516 9.17631 3.6094 9.06793 3.89658 9.02136C4.04493 9.0039 4.19567 9.02105 4.33603 9.07215C4.61331 9.16908 4.83184 9.38643 4.93076 9.66297C5.03905 9.96346 4.99607 10.5347 5.03037 10.8768C5.15593 12.1285 5.5318 13.3453 6.254 14.3856C8.21912 17.2126 12.5465 17.6029 15.505 16.2762C17.7033 15.2904 18.876 12.9902 18.9737 10.66C19.0011 10.3277 18.9782 9.95865 19.0899 9.64246ZM11.6446 2.02136C13.3014 1.8719 15.0753 2.55335 16.0636 3.90711C17.3039 5.60658 16.9297 7.57787 16.9933 9.51258C17.094 12.5795 15.5804 14.6964 12.4269 14.9735C11.0254 15.0315 9.70841 14.768 8.61631 13.829C7.65801 13.0048 7.11544 11.7587 7.04014 10.5096C6.99793 9.80921 7.02898 9.06921 7.02061 8.36316C6.99942 6.57705 6.92635 5.08207 8.18369 3.62488C9.06989 2.59789 10.3004 2.10617 11.6446 2.02136Z"
+      fill="#604903"
+    />
+  </svg>
+);
+MicIconPapole.displayName = 'MicIconPapole';
