+++ src/components/PapoleIcons/MicOffIconPapole.tsx
@@ -1,0 +1,27 @@
+import type { FC } from 'react';
+import type { PapoleIconProps } from './types';
+
+/**
+ * Microfone desligado — tema Papolê. Cor marrom `#604903` (fixa da arte).
+ * `aria-hidden` por ser decorativo; quem usa em botão deve prover o rótulo.
+ */
+export const MicOffIconPapole: FC<PapoleIconProps> = ({
+  size = 24,
+  className,
+}) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    className={className}
+    aria-hidden="true"
+  >
+    <path
+      d="M2.92138 2.00096C3.51243 1.97398 3.79735 2.38941 4.18114 2.77342L5.24071 3.83787L6.43603 5.03221C6.65214 5.24803 6.95078 5.53088 7.14013 5.75877C7.41612 4.73636 7.69377 4.07412 8.47899 3.31639C10.3488 1.51221 13.8942 1.57682 15.686 3.45994C16.928 4.74071 17.0078 6.30724 16.9936 7.9658C16.9725 10.4436 17.2409 12.039 15.2905 13.9053C15.4504 14.0281 15.7475 14.3379 15.897 14.4892C16.2051 14.7986 16.5109 15.1115 16.813 15.4267C18.1001 14.1651 18.7625 12.8312 18.9448 11.0283C19.0123 10.3609 18.8665 9.42655 19.5864 9.09959C19.8254 8.99225 20.0977 8.98518 20.3423 9.07908C20.5949 9.17395 20.7975 9.3682 20.9038 9.61619C21.022 9.89823 20.9779 10.3016 20.9614 10.6054C20.8956 11.8206 20.6659 12.8698 20.2183 13.999C19.8981 14.6879 19.5623 15.3325 19.0698 15.9189C18.8077 16.231 18.5097 16.5316 18.2339 16.8369C18.4118 16.9867 18.6925 17.2869 18.8696 17.4648L20.06 18.6582L21.1362 19.7285C21.6674 20.2521 22.3111 20.706 21.8325 21.541C21.6184 21.9356 20.8757 22.1607 20.5151 21.8642C19.9124 21.3681 19.3263 20.7287 18.7759 20.1748L15.3921 16.792L7.33251 8.7324L4.19774 5.59861C3.58409 4.98783 2.96253 4.37668 2.35497 3.75975C2.18785 3.59003 2.05404 3.38354 2.01317 3.14451C1.9702 2.89156 2.0312 2.63133 2.18212 2.42381C2.37884 2.15194 2.60825 2.05338 2.92138 2.00096ZM4.17724 9.09861C4.35579 9.08613 4.56672 9.14867 4.72313 9.2324C5.3088 9.54598 5.20095 10.4279 5.24853 11.0058C5.29609 11.546 5.39785 12.08 5.55224 12.5986C6.34419 15.2155 8.29261 16.6266 10.8843 16.9746C11.3668 17.0394 12.0198 16.9345 12.4526 17.1758C12.6687 17.2964 12.8225 17.4863 12.8872 17.7304C12.9826 18.0903 12.9844 20.6384 12.938 21.0605C12.9203 21.2214 12.8826 21.3792 12.8032 21.5205C12.6402 21.8103 12.3689 21.9179 12.0708 21.998C11.8022 22.0146 11.5208 21.9284 11.3276 21.7314C11.1975 21.5978 11.1064 21.4285 11.0659 21.2441C11.0016 20.9629 11.0318 19.3686 11.0308 18.9804C9.4007 18.7722 8.13113 18.4599 6.73388 17.4775C4.4657 15.8888 3.33397 13.1768 3.29052 10.415C3.27955 9.71664 3.39629 9.22846 4.17724 9.09861Z"
+      fill="#604903"
+    />
+  </svg>
+);
+MicOffIconPapole.displayName = 'MicOffIconPapole';
