+++ src/components/PapoleIcons/PauseIconPapole.tsx
@@ -1,111 +1,11 @@
 import { useId, type FC } from 'react';
-
-export interface PapoleIconProps {
-  /** Largura e altura em pixels. Default: tamanho nativo do ícone. */
-  size?: number;
-  /** Classes aplicadas ao `<svg>` raiz. */
-  className?: string;
-}
+import type { PapoleIconProps } from './types';
 
 /**
- * Ícones do tema Papolê (convertidos de `src/assets/icons/papole/*.svg`).
- *
- * As cores são as da arte (dourado `#F9CB3B` / contorno creme `#FDEFC4` para
- * play/pause; marrom `#604903` para microfone/stop). São `aria-hidden` por serem
- * decorativos — quem usa em botão deve prover o rótulo (ex.: `aria-label`).
+ * Pausar (duas barras douradas com contorno creme) — tema Papolê. Cores fixas da
+ * arte (`#F9CB3B` + `#FDEFC4`). O contorno usa um `mask` com id único (via
+ * `useId`) pra não colidir entre instâncias. `aria-hidden`; rótulo vem de quem usa.
  */
-
-/** Microfone (ligado). */
-export const MicIconPapole: FC<PapoleIconProps> = ({
-  size = 24,
-  className,
-}) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    className={className}
-    aria-hidden="true"
-  >
-    <path
-      d="M19.0899 9.64246C19.2722 9.12736 19.8996 8.87431 20.3966 9.09265C21.0667 9.38732 21.0012 10.0389 20.9688 10.6317C20.902 11.8547 20.6668 12.9383 20.1905 14.0663C19.8435 14.8412 19.4182 15.5484 18.8468 16.1815C17.3308 17.8298 15.2311 18.8217 12.9952 18.9471L13.0011 20.2665C13.0021 20.5743 13.0211 20.8906 12.9669 21.1962C12.9158 21.6151 12.4185 22.0384 11.9796 21.996C10.9121 21.8926 11.0195 20.9985 11.0206 20.3046L11.0235 18.955C6.82821 18.5619 3.81573 15.9926 3.14951 11.745C3.07466 11.2676 2.95457 10.2389 3.04502 9.78601C3.07111 9.6531 3.12649 9.52768 3.20713 9.41883C3.38516 9.17631 3.6094 9.06793 3.89658 9.02136C4.04493 9.0039 4.19567 9.02105 4.33603 9.07215C4.61331 9.16908 4.83184 9.38643 4.93076 9.66297C5.03905 9.96346 4.99607 10.5347 5.03037 10.8768C5.15593 12.1285 5.5318 13.3453 6.254 14.3856C8.21912 17.2126 12.5465 17.6029 15.505 16.2762C17.7033 15.2904 18.876 12.9902 18.9737 10.66C19.0011 10.3277 18.9782 9.95865 19.0899 9.64246ZM11.6446 2.02136C13.3014 1.8719 15.0753 2.55335 16.0636 3.90711C17.3039 5.60658 16.9297 7.57787 16.9933 9.51258C17.094 12.5795 15.5804 14.6964 12.4269 14.9735C11.0254 15.0315 9.70841 14.768 8.61631 13.829C7.65801 13.0048 7.11544 11.7587 7.04014 10.5096C6.99793 9.80921 7.02898 9.06921 7.02061 8.36316C6.99942 6.57705 6.92635 5.08207 8.18369 3.62488C9.06989 2.59789 10.3004 2.10617 11.6446 2.02136Z"
-      fill="#604903"
-    />
-  </svg>
-);
-MicIconPapole.displayName = 'MicIconPapole';
-
-/** Microfone desligado. */
-export const MicOffIconPapole: FC<PapoleIconProps> = ({
-  size = 24,
-  className,
-}) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    className={className}
-    aria-hidden="true"
-  >
-    <path
-      d="M2.92138 2.00096C3.51243 1.97398 3.79735 2.38941 4.18114 2.77342L5.24071 3.83787L6.43603 5.03221C6.65214 5.24803 6.95078 5.53088 7.14013 5.75877C7.41612 4.73636 7.69377 4.07412 8.47899 3.31639C10.3488 1.51221 13.8942 1.57682 15.686 3.45994C16.928 4.74071 17.0078 6.30724 16.9936 7.9658C16.9725 10.4436 17.2409 12.039 15.2905 13.9053C15.4504 14.0281 15.7475 14.3379 15.897 14.4892C16.2051 14.7986 16.5109 15.1115 16.813 15.4267C18.1001 14.1651 18.7625 12.8312 18.9448 11.0283C19.0123 10.3609 18.8665 9.42655 19.5864 9.09959C19.8254 8.99225 20.0977 8.98518 20.3423 9.07908C20.5949 9.17395 20.7975 9.3682 20.9038 9.61619C21.022 9.89823 20.9779 10.3016 20.9614 10.6054C20.8956 11.8206 20.6659 12.8698 20.2183 13.999C19.8981 14.6879 19.5623 15.3325 19.0698 15.9189C18.8077 16.231 18.5097 16.5316 18.2339 16.8369C18.4118 16.9867 18.6925 17.2869 18.8696 17.4648L20.06 18.6582L21.1362 19.7285C21.6674 20.2521 22.3111 20.706 21.8325 21.541C21.6184 21.9356 20.8757 22.1607 20.5151 21.8642C19.9124 21.3681 19.3263 20.7287 18.7759 20.1748L15.3921 16.792L7.33251 8.7324L4.19774 5.59861C3.58409 4.98783 2.96253 4.37668 2.35497 3.75975C2.18785 3.59003 2.05404 3.38354 2.01317 3.14451C1.9702 2.89156 2.0312 2.63133 2.18212 2.42381C2.37884 2.15194 2.60825 2.05338 2.92138 2.00096ZM4.17724 9.09861C4.35579 9.08613 4.56672 9.14867 4.72313 9.2324C5.3088 9.54598 5.20095 10.4279 5.24853 11.0058C5.29609 11.546 5.39785 12.08 5.55224 12.5986C6.34419 15.2155 8.29261 16.6266 10.8843 16.9746C11.3668 17.0394 12.0198 16.9345 12.4526 17.1758C12.6687 17.2964 12.8225 17.4863 12.8872 17.7304C12.9826 18.0903 12.9844 20.6384 12.938 21.0605C12.9203 21.2214 12.8826 21.3792 12.8032 21.5205C12.6402 21.8103 12.3689 21.9179 12.0708 21.998C11.8022 22.0146 11.5208 21.9284 11.3276 21.7314C11.1975 21.5978 11.1064 21.4285 11.0659 21.2441C11.0016 20.9629 11.0318 19.3686 11.0308 18.9804C9.4007 18.7722 8.13113 18.4599 6.73388 17.4775C4.4657 15.8888 3.33397 13.1768 3.29052 10.415C3.27955 9.71664 3.39629 9.22846 4.17724 9.09861Z"
-      fill="#604903"
-    />
-  </svg>
-);
-MicOffIconPapole.displayName = 'MicOffIconPapole';
-
-/** Parar (quadrado arredondado). */
-export const StopIconPapole: FC<PapoleIconProps> = ({
-  size = 24,
-  className,
-}) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    className={className}
-    aria-hidden="true"
-  >
-    <path
-      d="M9.19373 3.02192C9.23924 3.01968 9.28477 3.01795 9.3303 3.01669C10.3859 2.98615 11.5022 3.00736 12.5695 3.00486C14.1845 3.02035 15.7972 2.94773 17.3862 3.2905C20.5781 3.97907 20.9146 6.66832 20.9808 9.41075C20.9978 10.5335 21.003 11.6565 20.996 12.7793C20.9942 14.8218 21.144 17.7223 19.7633 19.3629C18.5786 20.7708 16.5405 20.9308 14.8126 20.9714C13.7196 21.0046 12.4391 21.0064 11.3323 20.9895C9.21141 20.9573 6.22451 21.2041 4.55526 19.6974C3.28925 18.5546 3.06985 16.632 3.02355 15.0153C2.99244 13.9292 2.99703 12.6416 3.00958 11.5439C3.03437 9.37514 2.76331 6.22534 4.32955 4.52291C5.55279 3.19327 7.51307 3.08323 9.19373 3.02192Z"
-      fill="#604903"
-    />
-  </svg>
-);
-StopIconPapole.displayName = 'StopIconPapole';
-
-/** Reproduzir (triângulo dourado com contorno creme). */
-export const PlayIconPapole: FC<PapoleIconProps> = ({
-  size = 30,
-  className,
-}) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={size}
-    height={size}
-    viewBox="0 0 30 30"
-    fill="none"
-    className={className}
-    aria-hidden="true"
-  >
-    <path
-      d="M9.69336 3.00977C10.4688 2.9555 11.2352 3.14459 11.7656 3.30371C12.3241 3.47126 12.8823 3.69117 13.2188 3.81738L13.2266 3.81934L13.2334 3.82227C16.6576 5.1362 19.8422 7.04365 22.6582 9.45996L23.2168 9.94922L23.2227 9.95508C23.672 10.3617 24.7284 11.2995 25.2793 12.2227L25.2871 12.2354L25.2939 12.2471C26.0279 13.5184 26.1156 14.9454 25.8896 16.1572L25.8408 16.3965C25.4502 18.1299 24.2837 19.1044 23.6602 19.6924C22.0107 21.2479 20.2264 22.4993 18.4365 23.6328L18.4248 23.6406C16.5421 24.8144 14.5418 25.7695 12.46 26.4893L12.4512 26.4932L12.4414 26.4961C12.05 26.6272 11.243 26.9066 10.4883 26.9756L10.4893 26.9766C10.4871 26.9768 10.4846 26.9764 10.4824 26.9766C10.4811 26.9767 10.4798 26.9774 10.4785 26.9775L10.4775 26.9766C9.01305 27.1114 7.57551 26.6212 6.48535 25.6455L6.48145 25.6416C6.03578 25.2405 5.5503 24.6967 5.20605 24.0059C4.77418 23.1392 4.5478 21.9412 4.41113 21.0146C4.26294 20.0099 4.17714 18.9939 4.12598 18.332L4.12695 18.3311C3.87395 15.2113 4.00141 12.0703 4.50684 8.98438L4.50781 8.97559C4.58505 8.51619 4.69474 7.82631 4.87891 7.15918C5.05634 6.51655 5.3614 5.66072 5.96387 4.95801C6.91568 3.8481 8.10898 3.14206 9.66699 3.01172L9.67969 3.01074L9.69336 3.00977Z"
-      fill="#F9CB3B"
-      stroke="#FDEFC4"
-      strokeWidth={4}
-    />
-  </svg>
-);
-PlayIconPapole.displayName = 'PlayIconPapole';
-
-/** Pausar (duas barras douradas com contorno creme). */
 export const PauseIconPapole: FC<PapoleIconProps> = ({
   size = 30,
   className,
