+++ src/components/PapoleIcons/PlayIconPapole.tsx
@@ -1,0 +1,29 @@
+import type { FC } from 'react';
+import type { PapoleIconProps } from './types';
+
+/**
+ * Reproduzir (triângulo dourado com contorno creme) — tema Papolê. Cores fixas
+ * da arte (`#F9CB3B` + `#FDEFC4`). `aria-hidden`; quem usa em botão dá o rótulo.
+ */
+export const PlayIconPapole: FC<PapoleIconProps> = ({
+  size = 30,
+  className,
+}) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size}
+    height={size}
+    viewBox="0 0 30 30"
+    fill="none"
+    className={className}
+    aria-hidden="true"
+  >
+    <path
+      d="M9.69336 3.00977C10.4688 2.9555 11.2352 3.14459 11.7656 3.30371C12.3241 3.47126 12.8823 3.69117 13.2188 3.81738L13.2266 3.81934L13.2334 3.82227C16.6576 5.1362 19.8422 7.04365 22.6582 9.45996L23.2168 9.94922L23.2227 9.95508C23.672 10.3617 24.7284 11.2995 25.2793 12.2227L25.2871 12.2354L25.2939 12.2471C26.0279 13.5184 26.1156 14.9454 25.8896 16.1572L25.8408 16.3965C25.4502 18.1299 24.2837 19.1044 23.6602 19.6924C22.0107 21.2479 20.2264 22.4993 18.4365 23.6328L18.4248 23.6406C16.5421 24.8144 14.5418 25.7695 12.46 26.4893L12.4512 26.4932L12.4414 26.4961C12.05 26.6272 11.243 26.9066 10.4883 26.9756L10.4893 26.9766C10.4871 26.9768 10.4846 26.9764 10.4824 26.9766C10.4811 26.9767 10.4798 26.9774 10.4785 26.9775L10.4775 26.9766C9.01305 27.1114 7.57551 26.6212 6.48535 25.6455L6.48145 25.6416C6.03578 25.2405 5.5503 24.6967 5.20605 24.0059C4.77418 23.1392 4.5478 21.9412 4.41113 21.0146C4.26294 20.0099 4.17714 18.9939 4.12598 18.332L4.12695 18.3311C3.87395 15.2113 4.00141 12.0703 4.50684 8.98438L4.50781 8.97559C4.58505 8.51619 4.69474 7.82631 4.87891 7.15918C5.05634 6.51655 5.3614 5.66072 5.96387 4.95801C6.91568 3.8481 8.10898 3.14206 9.66699 3.01172L9.67969 3.01074L9.69336 3.00977Z"
+      fill="#F9CB3B"
+      stroke="#FDEFC4"
+      strokeWidth={4}
+    />
+  </svg>
+);
+PlayIconPapole.displayName = 'PlayIconPapole';
