+++ src/components/PapoleIcons/index.ts
@@ -1,0 +1,6 @@
+export { MicIconPapole } from './MicIconPapole';
+export { MicOffIconPapole } from './MicOffIconPapole';
+export { StopIconPapole } from './StopIconPapole';
+export { PlayIconPapole } from './PlayIconPapole';
+export { PauseIconPapole } from './PauseIconPapole';
+export type { PapoleIconProps } from './types';
