+++ src/components/PapoleIcons/types.ts
@@ -1,0 +1,6 @@
+export interface PapoleIconProps {
+  /** Largura e altura em pixels. Default: tamanho nativo do ícone. */
+  size?: number;
+  /** Classes aplicadas ao `<svg>` raiz. */
+  className?: string;
+}
